Temporary fix for the Expired dropdown issue
The problem is that the showDropdownOptionJump method uses the ==
comparison operator instead of ===, so 0 technically equals "LIVE",
and Expired gets selected in the dropdown menu by default, since it
overrides the actual "LIVE" option.

It would be easy to just change the == operator to ===, however this
may break other instances of the method, so to fix this properly it's
going to take some time.

This issue exists on both the main Domains page as well as the main
SSL page.